+++ user-rest-service/interfaces/handler/user.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/apierrors"
-	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/infrastructure/response"
+	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/interfaces/presenter"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/usecase"
 	"github.com/paypay3/kakeibo-rest-api-ddd/user-rest-service/usecase/input"
 )
@@ -23,15 +23,15 @@ func NewUserHandler(userUsecase usecase.UserUsecase) *userHandler {
 func (h *userHandler) SignUp(w http.ResponseWriter, r *http.Request) {
 	var in input.SignUpUser
 	if err := json.NewDecoder(r.Body).Decode(&in); err != nil {
-		response.ErrorJSON(w, apierrors.NewBadRequestError(apierrors.NewErrorString("正しいデータを入力してください")))
+		presenter.ErrorJSON(w, apierrors.NewBadRequestError(apierrors.NewErrorString("正しいデータを入力してください")))
 		return
 	}
 
 	out, err := h.userUsecase.SignUp(&in)
 	if err != nil {
-		response.ErrorJSON(w, err)
+		presenter.ErrorJSON(w, err)
 		return
 	}
 
-	response.JSON(w, http.StatusCreated, out)
+	presenter.JSON(w, http.StatusCreated, out)
 }
