+++ user-rest-service/interfaces/presenter/http_success.go
@@ -1,4 +1,4 @@
-package response
+package presenter
 
 import (
 	"encoding/json"
